Move copy functions to their own file in prim/fun and have GPU types only work with doubles
